#1 refactor/qrcode rename file
